+++ changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### Version 1.1.0:
+- Cut to LinkBuilder.from(String) and LinkBuilder.on(TextView) to create an instance of LinkBuilder
+- Allow LinkBuilder to return a CharSequence for the links without setting them to a TextView automatically. (Use LinkBuilder.from(String).build())
+
 ### Version 1.0.6:
 - Allow for styling the default color of the links from activity theme
 - Don't allow null links when building the SpannableString